Move pickers to `lab` and delete package
