--- .github/workflows/formik-material-ui-pickers.yaml
@@ -1,33 +1,0 @@
-name: Build formik-material-ui
-on:
-  push:
-    branches:
-      - master
-  pull_request:
-    branches:
-      - '**'
-
-jobs:
-  build:
-    runs-on: ubuntu-latest
-
-    strategy:
-      matrix:
-        node-version: [14.x, 16.x]
-
-    steps:
-      - uses: actions/checkout@v1
-      - name: Use Node.js ${{ matrix.node-version }}
-        uses: actions/setup-node@v1
-        with:
-          node-version: ${{ matrix.node-version }}
-      - name: build
-        run: |
-          yarn install --frozen-lockfile
-          yarn typecheck
-          yarn lint
-          yarn test
-          yarn build
-        working-directory: packages/formik-material-ui-pickers
-        env:
-          CI: true
