+++ packages/formik-material-ui-lab/rollup.config.js
@@ -1,7 +1,24 @@
 import typescript from 'rollup-plugin-typescript2';
-import pkg from './package.json';
+import pkg from '../formik-material-ui-lab/package.json';
 
-const external = ['react', 'react-native', 'formik'];
+const external = [
+  'react',
+  'react-native',
+  'formik',
+  '@mui/lab/DatePicker',
+  '@mui/lab/DateTimePicker',
+  '@mui/lab/DesktopDatePicker',
+  '@mui/lab/DesktopDateTimePicker',
+  '@mui/lab/DesktopTimePicker',
+  '@mui/lab/MobileDatePicker',
+  '@mui/lab/MobileDateTimePicker',
+  '@mui/lab/MobileTimePicker',
+  '@mui/lab/StaticDatePicker',
+  '@mui/lab/StaticDateTimePicker',
+  '@mui/lab/StaticTimePicker',
+  '@mui/lab/TimePicker',
+  '@mui/material/TextField',
+];
 
 const config = [
   {
