+++ packages/formik-material-ui-lab/src/StaticDatePicker.tsx
