+++ packages/formik-material-ui-lab/src/StaticDateTimePicker.tsx
