+++ packages/formik-material-ui-lab/src/__tests__/StaticDatePicker.test.tsx
