+++ packages/formik-material-ui-lab/src/__tests__/StaticDateTimePicker.test.tsx
