+++ packages/formik-material-ui-lab/tsconfig.package.json
@@ -1,5 +1,5 @@
 {
   "extends": "../../tsconfig.json",
-  "include": ["src"],
+  "include": ["src", "types"],
   "exclude": ["node_modules", "dist", "**/__tests__/**"]
 }
