--- packages/formik-material-ui-pickers/README.md
@@ -1,10 +1,0 @@
-# Formik Material-UI Pickers
-
-![](https://github.com/stackworx/formik-material-ui/workflows/Build%20formik-material-ui-pickers/badge.svg)[![license](https://badgen.now.sh/badge/license/MIT)](./LICENSE)
-
-Bindings for using [Formik](https://github.com/jaredpalmer/formik) with [Material-UI Pickers](https://material-ui-pickers.dev/).
-
-- [Documentation](https://stackworx.github.io/formik-material-ui)
-- [Code Sandbox](https://codesandbox.io/s/915qlr56rp)
-
-This project requires Formik>= 2.0.0
