--- packages/formik-material-ui-pickers/package.json
@@ -1,66 +1,0 @@
-{
-  "name": "formik-material-ui-pickers",
-  "version": "0.0.12",
-  "license": "MIT",
-  "typings": "dist/main.d.ts",
-  "peerDependencies": {
-    "@emotion/react": "^11.4.1",
-    "@emotion/styled": "^11.3.0",
-    "@mui/lab": "^5.0.0-alpha.48",
-    "@mui/material": "^5.0.1",
-    "formik": ">=2.0.0",
-    "react": ">=16.8.0"
-  },
-  "keywords": [
-    "react",
-    "formik",
-    "material-ui",
-    "mui",
-    "form",
-    "pickers"
-  ],
-  "prettier": {
-    "singleQuote": true,
-    "trailingComma": "es5"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/stackworx/formik-material-ui"
-  },
-  "jest": {
-    "testURL": "http://localhost/",
-    "testEnvironment": "jsdom",
-    "transform": {
-      "^.+\\.tsx?$": "ts-jest"
-    },
-    "roots": [
-      "src"
-    ],
-    "testRegex": "(/src/__tests__/.*|(\\.|/)(test|spec))\\.(jsx?|tsx?)$",
-    "moduleFileExtensions": [
-      "ts",
-      "tsx",
-      "js",
-      "jsx",
-      "json",
-      "node"
-    ],
-    "globalSetup": "./global-setup.js"
-  },
-  "files": [
-    "dist"
-  ],
-  "main": "dist/index.js",
-  "module": "dist/formik-material-ui-pickers.es6.js",
-  "scripts": {
-    "typecheck": "tsc",
-    "build": "cross-env NODE_ENV=production tsc --noEmit --project ./tsconfig.package.json",
-    "package": "rimraf build dist && cross-env NODE_ENV=production tsc -p ./tsconfig.package.json && rollup -c && rimraf .rpt2_cache dist/.rpt2_cache",
-    "test": "jest",
-    "test:updateSnapshot": "jest --updateSnapshot",
-    "lint": "eslint \"./{src,stories}/**/*.{tsx,ts,js}\"",
-    "lint:fix": "eslint --fix \"./{src,stories}/**/*.{tsx,ts,js}\"",
-    "prepublishOnly": "yarn test && yarn build && yarn package"
-  },
-  "gitHead": "409cd27bdca327979c0dd7e20edd3ab773620396"
-}
