--- packages/formik-material-ui-pickers/src/main.ts
@@ -1,12 +1,0 @@
-export * from './DatePicker';
-export * from './DateTimePicker';
-export * from './DesktopDatePicker';
-export * from './DesktopDateTimePicker';
-export * from './DesktopTimePicker';
-export * from './MobileDatePicker';
-export * from './MobileDateTimePicker';
-export * from './MobileTimePicker';
-export * from './StaticDatePicker';
-export * from './StaticDateTimePicker';
-export * from './StaticTimePicker';
-export * from './TimePicker';
