--- packages/formik-material-ui-pickers/tsconfig.package.json
@@ -1,5 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "include": ["src", "types"],
-  "exclude": ["node_modules", "dist", "**/__tests__/**"]
-}
