+++ packages/formik-material-ui/README.md
@@ -1,7 +1,6 @@
 # Formik Material-UI
 
 ![](https://github.com/stackworx/formik-material-ui/workflows/Build%20formik-material-ui/badge.svg)
-![](https://github.com/stackworx/formik-material-ui/workflows/Build%20formik-material-ui-pickers/badge.svg)
 ![](https://github.com/stackworx/formik-material-ui/workflows/Build%20formik-material-ui-lab/badge.svg)
 [![license](https://badgen.now.sh/badge/license/MIT)](./LICENSE)
